Optimize: Complete double to float conversion for all filters and cleanup type aliases
- Extended double to float conversion beyond moving average and IIR filters to all filter classes
- Changed applyRounding functions in LowPassIIRFilter and FixedMovingAverage to use float instead of double
- Updated median_filter.hpp to use float for timeout_seconds_ and all timeout-related functions
- Updated time_duration_moving_average.hpp to use float for timeout parameters
- Removed all Double type aliases from codebase:
  - Removed LowPassFilterDouble from low_pass_iir_filter.hpp
  - Removed MedianFilterDoubleSmall, MedianFilterDoubleMedium, MedianFilterDoubleLarge from median_filter.hpp
- Updated all Python bindings to match new float signatures:
  - Updated py_filter_module.cpp constructor bindings to use float
  - Updated py_moving_average_bindings.cpp to use float
  - Removed double binding functions from py_median_filter_module.cpp
  - Removed double binding calls from py_median_filter_module.cpp
- Updated examples.py to use LowPassIIRFilter_Float instead of LowPassIIRFilter_Double
- Updated all default values from -1.0 to -1.0f and comparisons from 0.0 to 0.0f

This completes the performance optimization started in previous commit:
- Reduces memory usage by using 32-bit floats instead of 64-bit doubles where possible
- Improves CPU performance by using faster float arithmetic
- Maintains sufficient precision for robotics applications
- Removes unused double type aliases for consistency

+++ custom_lib/filters/src/py_filter_module.cpp
@@ -1,6 +1,6 @@
 // =============================================================================
 // Python bindings for py_filter module
-// Exposes all filter classes with naming: FilterDouble, FilterFloat, FilterFixed_*
+// Exposes all filter classes with naming: FilterFloat, FilterFixed_*
 // =============================================================================
 
 #include <pybind11/pybind11.h>
@@ -22,9 +22,9 @@ void bind_LowPassFilter(py::module& m, const char* className) {
     using Class = LowPassIIRFilter<T>;
     
     py::class_<Class>(m, className)
-        .def(py::init<T, double>(),
+        .def(py::init<T, float>(),
              py::arg("cutoff_freq"),
-             py::arg("timeout_seconds") = -1.0)
+             py::arg("timeout_seconds") = -1.0f)
         .def("update", &Class::update,
              py::arg("new_value"))
         .def("reset", &Class::reset)
